+++ src/gadgets/boolean.rs
@@ -39,9 +39,9 @@ impl AllocatedBit {
             || "boolean",
             || {
                 if *value.get()? {
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -76,9 +76,9 @@ impl AllocatedBit {
             || "boolean",
             || {
                 if *value.get()? {
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -113,11 +113,11 @@ impl AllocatedBit {
                 if *a.value.get()? ^ *b.value.get()? {
                     result_value = Some(true);
 
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
                     result_value = Some(false);
 
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -165,11 +165,11 @@ impl AllocatedBit {
                 if *a.value.get()? & *b.value.get()? {
                     result_value = Some(true);
 
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
                     result_value = Some(false);
 
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -203,11 +203,11 @@ impl AllocatedBit {
                 if *a.value.get()? & !*b.value.get()? {
                     result_value = Some(true);
 
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
                     result_value = Some(false);
 
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -241,11 +241,11 @@ impl AllocatedBit {
                 if !*a.value.get()? & !*b.value.get()? {
                     result_value = Some(true);
 
-                    Ok(Scalar::one())
+                    Ok(Scalar::ONE)
                 } else {
                     result_value = Some(false);
 
-                    Ok(Scalar::zero())
+                    Ok(Scalar::ZERO)
                 }
             },
         )?;
@@ -390,7 +390,7 @@ impl Boolean {
                     || "enforce equal to one",
                     |lc| lc,
                     |lc| lc,
-                    |lc| lc + CS::one() - &a.lc(CS::one(), Scalar::one()),
+                    |lc| lc + CS::one() - &a.lc(CS::one(), Scalar::ONE),
                 );
 
                 Ok(())
@@ -400,7 +400,7 @@ impl Boolean {
                     || "enforce equal to zero",
                     |lc| lc,
                     |lc| lc,
-                    |_| a.lc(CS::one(), Scalar::one()),
+                    |_| a.lc(CS::one(), Scalar::ONE),
                 );
 
                 Ok(())
@@ -410,7 +410,7 @@ impl Boolean {
                     || "enforce equal",
                     |lc| lc,
                     |lc| lc,
-                    |_| a.lc(CS::one(), Scalar::one()) - &b.lc(CS::one(), Scalar::one()),
+                    |_| a.lc(CS::one(), Scalar::ONE) - &b.lc(CS::one(), Scalar::ONE),
                 );
 
                 Ok(())
@@ -599,16 +599,16 @@ impl Boolean {
             || {
                 ch_value
                     .get()
-                    .map(|v| if *v { Scalar::one() } else { Scalar::zero() })
+                    .map(|v| if *v { Scalar::ONE } else { Scalar::ZERO })
             },
         )?;
 
         // a(b - c) = ch - c
         cs.enforce(
             || "ch computation",
-            |_| b.lc(CS::one(), Scalar::one()) - &c.lc(CS::one(), Scalar::one()),
-            |_| a.lc(CS::one(), Scalar::one()),
-            |lc| lc + ch - &c.lc(CS::one(), Scalar::one()),
+            |_| b.lc(CS::one(), Scalar::ONE) - &c.lc(CS::one(), Scalar::ONE),
+            |_| a.lc(CS::one(), Scalar::ONE),
+            |lc| lc + ch - &c.lc(CS::one(), Scalar::ONE),
         );
 
         Ok(AllocatedBit {
@@ -702,7 +702,7 @@ impl Boolean {
             || {
                 maj_value
                     .get()
-                    .map(|v| if *v { Scalar::one() } else { Scalar::zero() })
+                    .map(|v| if *v { Scalar::ONE } else { Scalar::ZERO })
             },
         )?;
 
@@ -720,12 +720,12 @@ impl Boolean {
         cs.enforce(
             || "maj computation",
             |_| {
-                bc.lc(CS::one(), Scalar::one()) + &bc.lc(CS::one(), Scalar::one())
-                    - &b.lc(CS::one(), Scalar::one())
-                    - &c.lc(CS::one(), Scalar::one())
+                bc.lc(CS::one(), Scalar::ONE) + &bc.lc(CS::one(), Scalar::ONE)
+                    - &b.lc(CS::one(), Scalar::ONE)
+                    - &c.lc(CS::one(), Scalar::ONE)
             },
-            |_| a.lc(CS::one(), Scalar::one()),
-            |_| bc.lc(CS::one(), Scalar::one()) - maj,
+            |_| a.lc(CS::one(), Scalar::ONE),
+            |_| bc.lc(CS::one(), Scalar::ONE) - maj,
         );
 
         Ok(AllocatedBit {
@@ -755,9 +755,9 @@ mod test {
         let mut cs = TestConstraintSystem::new();
 
         AllocatedBit::alloc(&mut cs, Some(true)).unwrap();
-        assert!(cs.get("boolean") == Scalar::one());
+        assert!(cs.get("boolean") == Scalar::ONE);
         assert!(cs.is_satisfied());
-        cs.set("boolean", Scalar::zero());
+        cs.set("boolean", Scalar::ZERO);
         assert!(cs.is_satisfied());
         cs.set("boolean", Scalar::from(2));
         assert!(!cs.is_satisfied());
@@ -775,14 +775,14 @@ mod test {
                 assert_eq!(c.value.unwrap(), *a_val ^ *b_val);
 
                 assert!(cs.is_satisfied());
-                assert!(cs.get("a/boolean") == if *a_val { Field::one() } else { Field::zero() });
-                assert!(cs.get("b/boolean") == if *b_val { Field::one() } else { Field::zero() });
+                assert!(cs.get("a/boolean") == if *a_val { Field::ONE } else { Field::ZERO });
+                assert!(cs.get("b/boolean") == if *b_val { Field::ONE } else { Field::ZERO });
                 assert!(
                     cs.get("xor result")
                         == if *a_val ^ *b_val {
-                            Field::one()
+                            Field::ONE
                         } else {
-                            Field::zero()
+                            Field::ZERO
                         }
                 );
 
@@ -790,9 +790,9 @@ mod test {
                 cs.set(
                     "xor result",
                     if *a_val ^ *b_val {
-                        Field::zero()
+                        Field::ZERO
                     } else {
-                        Field::one()
+                        Field::ONE
                     },
                 );
                 assert!(!cs.is_satisfied());
@@ -811,14 +811,14 @@ mod test {
                 assert_eq!(c.value.unwrap(), *a_val & *b_val);
 
                 assert!(cs.is_satisfied());
-                assert!(cs.get("a/boolean") == if *a_val { Field::one() } else { Field::zero() });
-                assert!(cs.get("b/boolean") == if *b_val { Field::one() } else { Field::zero() });
+                assert!(cs.get("a/boolean") == if *a_val { Field::ONE } else { Field::ZERO });
+                assert!(cs.get("b/boolean") == if *b_val { Field::ONE } else { Field::ZERO });
                 assert!(
                     cs.get("and result")
                         == if *a_val & *b_val {
-                            Field::one()
+                            Field::ONE
                         } else {
-                            Field::zero()
+                            Field::ZERO
                         }
                 );
 
@@ -826,9 +826,9 @@ mod test {
                 cs.set(
                     "and result",
                     if *a_val & *b_val {
-                        Field::zero()
+                        Field::ZERO
                     } else {
-                        Field::one()
+                        Field::ONE
                     },
                 );
                 assert!(!cs.is_satisfied());
@@ -847,14 +847,14 @@ mod test {
                 assert_eq!(c.value.unwrap(), *a_val & !*b_val);
 
                 assert!(cs.is_satisfied());
-                assert!(cs.get("a/boolean") == if *a_val { Field::one() } else { Field::zero() });
-                assert!(cs.get("b/boolean") == if *b_val { Field::one() } else { Field::zero() });
+                assert!(cs.get("a/boolean") == if *a_val { Field::ONE } else { Field::ZERO });
+                assert!(cs.get("b/boolean") == if *b_val { Field::ONE } else { Field::ZERO });
                 assert!(
                     cs.get("and not result")
                         == if *a_val & !*b_val {
-                            Field::one()
+                            Field::ONE
                         } else {
-                            Field::zero()
+                            Field::ZERO
                         }
                 );
 
@@ -862,9 +862,9 @@ mod test {
                 cs.set(
                     "and not result",
                     if *a_val & !*b_val {
-                        Field::zero()
+                        Field::ZERO
                     } else {
-                        Field::one()
+                        Field::ONE
                     },
                 );
                 assert!(!cs.is_satisfied());
@@ -883,14 +883,14 @@ mod test {
                 assert_eq!(c.value.unwrap(), !*a_val & !*b_val);
 
                 assert!(cs.is_satisfied());
-                assert!(cs.get("a/boolean") == if *a_val { Field::one() } else { Field::zero() });
-                assert!(cs.get("b/boolean") == if *b_val { Field::one() } else { Field::zero() });
+                assert!(cs.get("a/boolean") == if *a_val { Field::ONE } else { Field::ZERO });
+                assert!(cs.get("b/boolean") == if *b_val { Field::ONE } else { Field::ZERO });
                 assert!(
                     cs.get("nor result")
                         == if !*a_val & !*b_val {
-                            Field::one()
+                            Field::ONE
                         } else {
-                            Field::zero()
+                            Field::ZERO
                         }
                 );
 
@@ -898,9 +898,9 @@ mod test {
                 cs.set(
                     "nor result",
                     if !*a_val & !*b_val {
-                        Field::zero()
+                        Field::ZERO
                     } else {
-                        Field::one()
+                        Field::ONE
                     },
                 );
                 assert!(!cs.is_satisfied());
@@ -1152,7 +1152,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1160,7 +1160,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1168,7 +1168,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1176,7 +1176,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
 
@@ -1187,7 +1187,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1195,7 +1195,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1203,7 +1203,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1211,7 +1211,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
 
@@ -1222,7 +1222,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1230,7 +1230,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1238,7 +1238,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1246,7 +1246,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
 
@@ -1257,7 +1257,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1265,7 +1265,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Not(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1273,7 +1273,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::one());
+                        assert!(cs.get("xor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1281,7 +1281,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("xor result") == Field::zero());
+                        assert!(cs.get("xor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
 
@@ -1370,7 +1370,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and result") == Field::one());
+                        assert!(cs.get("and result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1378,7 +1378,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and result") == Field::zero());
+                        assert!(cs.get("and result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1386,7 +1386,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1394,7 +1394,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::one());
+                        assert!(cs.get("and not result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
 
@@ -1406,7 +1406,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and result") == Field::zero());
+                        assert!(cs.get("and result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1414,7 +1414,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and result") == Field::zero());
+                        assert!(cs.get("and result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1422,7 +1422,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1430,7 +1430,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
 
@@ -1445,7 +1445,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1453,7 +1453,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1461,7 +1461,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("nor result") == Field::zero());
+                        assert!(cs.get("nor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1469,7 +1469,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("nor result") == Field::zero());
+                        assert!(cs.get("nor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
 
@@ -1484,7 +1484,7 @@ mod test {
                         OperandType::AllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::one());
+                        assert!(cs.get("and not result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
                     (
@@ -1492,7 +1492,7 @@ mod test {
                         OperandType::AllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("and not result") == Field::zero());
+                        assert!(cs.get("and not result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1500,7 +1500,7 @@ mod test {
                         OperandType::NegatedAllocatedTrue,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("nor result") == Field::zero());
+                        assert!(cs.get("nor result") == Field::ZERO);
                         assert_eq!(v.value, Some(false));
                     }
                     (
@@ -1508,7 +1508,7 @@ mod test {
                         OperandType::NegatedAllocatedFalse,
                         Boolean::Is(ref v),
                     ) => {
-                        assert!(cs.get("nor result") == Field::one());
+                        assert!(cs.get("nor result") == Field::ONE);
                         assert_eq!(v.value, Some(true));
                     }
 
@@ -1641,16 +1641,16 @@ mod test {
                     } else {
                         assert_eq!(cs.get("ch"), {
                             if expected {
-                                Scalar::one()
+                                Scalar::ONE
                             } else {
-                                Scalar::zero()
+                                Scalar::ZERO
                             }
                         });
                         cs.set("ch", {
                             if expected {
-                                Scalar::zero()
+                                Scalar::ZERO
                             } else {
-                                Scalar::one()
+                                Scalar::ONE
                             }
                         });
                         assert_eq!(cs.which_is_unsatisfied().unwrap(), "ch computation");
@@ -1733,16 +1733,16 @@ mod test {
                     } else {
                         assert_eq!(cs.get("maj"), {
                             if expected {
-                                Scalar::one()
+                                Scalar::ONE
                             } else {
-                                Scalar::zero()
+                                Scalar::ZERO
                             }
                         });
                         cs.set("maj", {
                             if expected {
-                                Scalar::zero()
+                                Scalar::ZERO
                             } else {
-                                Scalar::one()
+                                Scalar::ONE
                             }
                         });
                         assert_eq!(cs.which_is_unsatisfied().unwrap(), "maj computation");