+++ src/gadgets/test/mod.rs
@@ -2,6 +2,9 @@
 
 use ff::PrimeField;
 
+#[cfg(test)]
+use ff::Field;
+
 use crate::{ConstraintSystem, Index, LinearCombination, SynthesisError, Variable};
 
 use std::collections::HashMap;
@@ -69,7 +72,7 @@ fn proc_lc<Scalar: PrimeField>(terms: &[(Variable, Scalar)]) -> BTreeMap<Ordered
     let mut map = BTreeMap::new();
     for &(var, coeff) in terms {
         map.entry(OrderedVariable(var))
-            .or_insert_with(Scalar::zero)
+            .or_insert(Scalar::ZERO)
             .add_assign(&coeff);
     }
 
@@ -122,7 +125,7 @@ fn eval_lc<Scalar: PrimeField>(
     inputs: &[(Scalar, String)],
     aux: &[(Scalar, String)],
 ) -> Scalar {
-    let mut acc = Scalar::zero();
+    let mut acc = Scalar::ZERO;
 
     for &(var, ref coeff) in terms {
         let mut tmp = match var.get_unchecked() {
@@ -155,7 +158,7 @@ impl<Scalar: PrimeField> TestConstraintSystem<Scalar> {
             named_objects: map,
             current_namespace: vec![],
             constraints: vec![],
-            inputs: vec![(Scalar::one(), "ONE".into())],
+            inputs: vec![(Scalar::ONE, "ONE".into())],
             aux: vec![],
         }
     }
@@ -163,7 +166,7 @@ impl<Scalar: PrimeField> TestConstraintSystem<Scalar> {
     pub fn pretty_print(&self) -> String {
         let mut s = String::new();
 
-        let negone = Scalar::one().neg();
+        let negone = Scalar::ONE.neg();
 
         let powers_of_two = (0..Scalar::NUM_BITS)
             .map(|i| Scalar::from(2).pow_vartime(&[u64::from(i)]))
@@ -180,7 +183,7 @@ impl<Scalar: PrimeField> TestConstraintSystem<Scalar> {
                 }
                 is_first = false;
 
-                if coeff != Scalar::one() && coeff != negone {
+                if coeff != Scalar::ONE && coeff != negone {
                     for (i, x) in powers_of_two.iter().enumerate() {
                         if x == &coeff {
                             write!(s, "2^{} . ", i).unwrap();
@@ -462,8 +465,8 @@ fn test_cs() {
     {
         let mut cs = cs.namespace(|| "test1");
         let mut cs = cs.namespace(|| "test2");
-        cs.alloc(|| "hehe", || Ok(Scalar::one())).unwrap();
+        cs.alloc(|| "hehe", || Ok(Scalar::ONE)).unwrap();
     }
 
-    assert!(cs.get("test1/test2/hehe") == Scalar::one());
+    assert!(cs.get("test1/test2/hehe") == Scalar::ONE);
 }